Implement security checks for flashcard deletion and posting (#57)
* Implement security checks for flashcard deletion and posting

This commit introduces essential security checks to restrict unauthorized actions related to flashcard deletion and posting functionalities. Users are now required to have the appropriate permissions before being allowed to delete or post flashcards.

Changes include:
- Added authorization logic for flashcard deletion to ensure that only the creator of a flashcard can delete it.
- Implemented access controls for flashcard posting, allowing only authenticated users with the necessary posting privileges to create new flashcards.

This security enhancement safeguards sensitive flashcard content and maintains data integrity within the application.

* implemented tests for newly added methods, as well as refactored previous tests to match refactored methods

+++ src/main/java/com/codersfactory/flashcards/FlashcardsService.java
@@ -3,6 +3,8 @@ package com.codersfactory.flashcards;
 import com.codersfactory.flashcards.dto.CreateFlashcardDto;
 import com.codersfactory.flashcards.dto.FlashcardDto;
 import com.codersfactory.flashcards.exception.EntityNotFoundException;
+import com.codersfactory.flashcards.exception.UnauthorizedAction;
+import com.codersfactory.flashcards.exception.UnauthorizedEditException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -30,11 +32,20 @@ public class FlashcardsService {
         return mapper.toDto(flashcard);
     }
 
+    FlashcardDto getDtoById(Long id) {
+        return mapEntity(findById(id));
+    }
+
     public void saveEntity(Flashcard flashcard) {
         repository.save(flashcard);
     }
 
-    public void deleteById(Long id) {
-        repository.deleteById(id);
+    public void deleteById(Long id, String username) {
+        if (checkIfUserIsOwner(id, username)) repository.deleteById(id);
+        else throw new UnauthorizedEditException(Flashcard.class, UnauthorizedAction.DELETE);
+    }
+
+    private boolean checkIfUserIsOwner(Long id, String username) {
+        return findById(id).getFlashcardCollection().getUser().getUsername().equals(username);
     }
 }
